requirements.txt
Django>=4.2
djangorestframework>=3.14
drf-spectacular>=0.26.1
psycopg2-binary>=2.9
python-dotenv>=1.0
dj-database-url>=1.0
drf-yasg>=1.21
